requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
rich>=13.0.0
questionary>=2.0.0
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-mock>=3.10.0
